fix flickering after signin
