+++ portato/src/App.css
@@ -44,6 +44,7 @@
   font-weight:bold;
 }
 
+
 @media(max-width: 576px) {
   .ant-picker-panels {
     flex-direction: column;
