+++ portato/src/Components/FirebaseAuth.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useRef } from "react";
 import "firebaseui/dist/firebaseui.css";
-import { auth } from "../firebaseConfig";
 import { GoogleAuthProvider, EmailAuthProvider } from "firebase/auth";
 import { uiInstance, uiConfig } from "./firebaseUIInstance";
