+++ portato/src/Pages/Layouts/PageLayoutTest.tsx
@@ -1,21 +1,18 @@
 import React, { ReactNode, useState } from "react";
-import AppHeader from "../../Components/HeaderBar";
-import ButtomHeaderBar from "../../Components/ButtonHeaderBar";
-import SideBarNav from "../../Components/SideBarNav";
 
 import { Layout } from "antd";
 import BottomBar from "../../Components/BottomBar";
-const { Header, Content, Footer } = Layout;
+const { Content } = Layout;
 
 interface Props {
   children: ReactNode;
+  display?: string;
 }
 
 const MainLayout = (props: Props) => {
-  const { children } = props;
-  const [openMenu, setOpenMenu] = useState(false);
+  const { children, display = "" } = props;
   return (
-    <Layout className="main-layout">
+    <Layout style={{ display: display }} className="main-layout">
       <Content className="layout-children">{children}</Content>
       <BottomBar />
     </Layout>
